+++ headers/Hitable.hpp
@@ -7,7 +7,7 @@
 class Material;
 
 struct HitRecord{
-	float t;				//distance traveled to hit point
+	double t;				//distance traveled to hit point
 	Vec3 p;					//hit point
 	Vec3 normal;			//surface normal  
 	Material *material; 	//material
@@ -16,7 +16,7 @@ struct HitRecord{
 class Hitable{
 public:
 	virtual ~Hitable(){}
-	virtual bool hit(const Ray& r, float t_min, float t_max, HitRecord& rec) const = 0;
+	virtual bool hit(const Ray& r, double t_min, double t_max, HitRecord& rec) const = 0;
 	virtual bool bounding_box(AABB& box) const = 0;
 };
